% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bh_competition.R
\name{bh_competition}
\alias{bh_competition}
\title{Beverton-Holt competition}
\usage{
bh_competition(
  r = 1.5,
  alpha = 1,
  initial_N = 10,
  timesteps = 30,
  discrete = FALSE,
  final_N = FALSE,
  extinction_thresh = 0.01
)
}
\arguments{
\item{r}{growth rate of species}

\item{alpha}{competition coefficients}

\item{initial_N}{initial abundances}

\item{timesteps}{length of time to simulate dynamics}

\item{discrete}{whether or not abundances should be discrete; discretization leads to a stochastic model}

\item{final_N}{whether or not function should return only the final N after all timesteps}
}
\description{
Beverton-Holt competition
}
