% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_emigration.R
\name{dd_emigration}
\alias{dd_emigration}
\title{Density-dependent emigration}
\usage{
dd_emigration(
  N,
  max = 0.3,
  a = 30,
  threshold = 0,
  discrete = FALSE,
  dd = FALSE
)
}
\arguments{
\item{N}{number of individuals}

\item{max}{saturation value of emigration probability}

\item{a}{nonlinearity parameter (i.e. half-saturation constant) that controls rate of increase}

\item{threshold}{minimum value of N in which emigration is possible}

\item{discrete}{whether or not to return emigrants as discrete individuals, setting discrete == TRUE also makes the model stochastic}

\item{dd}{whether or not to implement density-dependence, if FALSE then a constant emigration rate (i.e. max rate)}
}
\description{
Density-dependent emigration
}
