% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborhood.R
\name{neighborhood}
\alias{neighborhood}
\title{Neighborhood}
\usage{
neighborhood(x, y, df, nh_size = 1, neighborhood = "vonNeumann", torus = TRUE)
}
\arguments{
\item{x, y}{coordinates of focal cell}

\item{df}{dataframe in which all the coordinates on the grid is stored}

\item{nh_size}{size of the neighborhood to be returned}

\item{neighborhood}{the type of neighborhood to use; vonNeumann by default, Moore otherwise}

\item{torus}{defined how the boundary condition should be treated; if TRUE, then the grid is toroidal}
}
\description{
Returns neighborhood coordinates given focal cell and neighborhood size
}
