% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_dynamics.R
\name{simulate_dynamics}
\alias{simulate_dynamics}
\title{Simulate metacommunity dynamics}
\usage{
simulate_dynamics(
  initial_df,
  r,
  alpha,
  delta,
  timesteps,
  disp_rate = 0.1,
  nh_size = 1,
  nh = "vonNeumann",
  dd_emi = FALSE,
  torus = TRUE,
  stochastic = FALSE
)
}
\arguments{
\item{initial_df}{dataframe specifying the initial conditions with ID, xy coordinates, speices denisities}

\item{r}{growth rate of species}

\item{alpha}{length(r)^2 competition matrix}

\item{delta}{aggregation parameter where 0 is uniform weighting, >0 high densities are weighted more, <0 less}

\item{timesteps}{length of time to simulate dynamics}

\item{disp_rate}{dispersal rate which can range from 0 (no dispersal) to 1 (maximum dispersal)}

\item{nh_size}{Manhattan distance specifying size of neighborhood}

\item{nh}{Type of neighborhood for dispersal to occur. Can be either 'Von Neummann' or 'Moore'}

\item{dd_emi}{If TRUE, emigration is density-dependent and follows type-II functional response}

\item{torus}{If TRUE, the lattice is wrapped around to remove edge effects}

\item{stochastic}{If TRUE, denisities are discretized and the process is stochastic}
}
\description{
Simulate metacommunity dynamics
}
